Move jet.identify_phase and jet.rate_of_climb_descent to flight module.
