+++ tests/unit/test_spire.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 import numpy as np
 import pandas as pd
 
-import pycontrails.datalib.spire.data_cleaning as spire
-from pycontrails.core import adsb
+from pycontrails.datalib.spire import data_cleaning
+from pycontrails.physics import jet
 
 from .conftest import get_static_path
 
@@ -14,7 +14,7 @@ from .conftest import get_static_path
 def test_noise_removal_cruise_altitude():
     """Check that noise in cruise altitude is removed."""
     altitude_ft = np.array([40000, 39975, 40000, 40000, 39975, 40000, 40000, 40025, 40025, 40000])
-    altitude_ft_cleaned = adsb.remove_noise_in_cruise_altitude(np.copy(altitude_ft))
+    altitude_ft_cleaned = data_cleaning.remove_noise_in_cruise_altitude(np.copy(altitude_ft))
     np.testing.assert_array_equal(altitude_ft_cleaned, 40000)
 
 
@@ -23,14 +23,14 @@ def test_separate_unique_flights_using_metadata():
     difference in flight metadata.
     """
     df_icao_address = pd.read_parquet(get_static_path("flight-spire-data-cleaning.pq"))
-    df_icao_address = adsb.downsample_waypoints(
+    df_icao_address = data_cleaning.downsample_waypoints(
         df_icao_address, time_resolution=10, time_var="timestamp"
     )
     df_icao_address = df_icao_address.astype({"on_ground": bool})
-    df_icao_address = spire._fill_missing_callsign_for_satellite_waypoints(df_icao_address)
+    df_icao_address = data_cleaning._fill_missing_callsign_for_satellite_waypoints(df_icao_address)
 
     df_test_1 = df_icao_address.copy()
-    flights = adsb.separate_unique_flights_from_waypoints(
+    flights = data_cleaning.separate_unique_flights_from_waypoints(
         df_test_1, columns=["tail_number", "aircraft_type_icao", "callsign"]
     )
     assert len(flights) == 4  # Identified four unique flights
@@ -41,11 +41,11 @@ def test_separate_unique_flights_using_ground_indicator():
     ground indicator.
     """
     df_icao_address = pd.read_parquet(get_static_path("flight-spire-data-cleaning.pq"))
-    df_icao_address = adsb.downsample_waypoints(
+    df_icao_address = data_cleaning.downsample_waypoints(
         df_icao_address, time_resolution=10, time_var="timestamp"
     )
     df_icao_address = df_icao_address.astype({"on_ground": bool})
-    df_icao_address = spire._fill_missing_callsign_for_satellite_waypoints(df_icao_address)
+    df_icao_address = data_cleaning._fill_missing_callsign_for_satellite_waypoints(df_icao_address)
 
     # Construct erroneous subset of waypoints consisting of two unique flights with the same callsign
     callsigns = ["SHT88J", "BAW506"]
@@ -54,23 +54,23 @@ def test_separate_unique_flights_using_ground_indicator():
     df_test_2["callsign"] = "killer-whale-1"
 
     # Unable to identify unique flights because metadata is the same
-    flights = adsb.separate_unique_flights_from_waypoints(
+    flights = data_cleaning.separate_unique_flights_from_waypoints(
         df_test_2, columns=["tail_number", "aircraft_type_icao", "callsign"]
     )
     assert len(flights) == 1
 
-    flights = spire.separate_flights_with_ground_indicator(flights)
+    flights = data_cleaning.separate_flights_with_ground_indicator(flights)
     assert len(flights) == 2  # Identified two unique flights
 
 
 def test_separate_unique_flights_with_multiple_cruise_phase():
     """Test algorithms to identify and separate unique flight trajectories with multiple cruise phases."""
     df_icao_address = pd.read_parquet(get_static_path("flight-spire-data-cleaning.pq"))
-    df_icao_address = adsb.downsample_waypoints(
+    df_icao_address = data_cleaning.downsample_waypoints(
         df_icao_address, time_resolution=10, time_var="timestamp"
     )
     df_icao_address = df_icao_address.astype({"on_ground": bool})
-    df_icao_address = spire._fill_missing_callsign_for_satellite_waypoints(df_icao_address)
+    df_icao_address = data_cleaning._fill_missing_callsign_for_satellite_waypoints(df_icao_address)
 
     # Construct erroneous subset of waypoints consisting of two unique flights
     callsigns = ["BAW506", "BAW507"]
@@ -86,39 +86,41 @@ def test_separate_unique_flights_with_multiple_cruise_phase():
     dt_sec = np.diff(df_test_3["timestamp"].values, append=np.datetime64("NaT")) / np.timedelta64(
         1, "s"
     )
-    flight_phase: adsb.FlightPhaseDetailed = adsb.identify_phase_of_flight_detailed(
+    flight_phase: data_cleaning.FlightPhaseDetailed = jet.identify_phase_of_flight_detailed(
         df_test_3["altitude_baro"].values, dt_sec, threshold_rocd=250, min_cruise_alt_ft=20000
     )
 
     # Identify multiple cruise phase
-    trajectory_check: adsb.TrajectoryCheck = adsb.identify_multiple_cruise_phases_and_diversion(
-        flight_phase,
-        df_test_3["longitude"].values,
-        df_test_3["latitude"].values,
-        df_test_3["altitude_baro"].values,
-        dt_sec,
-        ground_indicator=df_test_3["on_ground"].values,
+    trajectory_check: data_cleaning.TrajectoryCheck = (
+        data_cleaning.identify_multiple_cruise_phases_and_diversion(
+            flight_phase,
+            df_test_3["longitude"].values,
+            df_test_3["latitude"].values,
+            df_test_3["altitude_baro"].values,
+            dt_sec,
+            ground_indicator=df_test_3["on_ground"].values,
+        )
     )
     assert trajectory_check.multiple_cruise_phase
 
     # Unable to identify unique flights because metadata is the same
-    flights = adsb.separate_unique_flights_from_waypoints(
+    flights = data_cleaning.separate_unique_flights_from_waypoints(
         df_test_3, columns=["tail_number", "aircraft_type_icao", "callsign"]
     )
     assert len(flights) == 1
 
-    flights = spire.separate_flights_with_multiple_cruise_phase(flights)
+    flights = data_cleaning.separate_flights_with_multiple_cruise_phase(flights)
     assert len(flights) == 2
 
 
 def test_identify_flight_diversion():
     """Test algorithms to identify flight diversion, and no separation is done."""
     df_icao_address = pd.read_parquet(get_static_path("flight-spire-data-cleaning.pq"))
-    df_icao_address = adsb.downsample_waypoints(
+    df_icao_address = data_cleaning.downsample_waypoints(
         df_icao_address, time_resolution=10, time_var="timestamp"
     )
     df_icao_address = df_icao_address.astype({"on_ground": bool})
-    df_icao_address = spire._fill_missing_callsign_for_satellite_waypoints(df_icao_address)
+    df_icao_address = data_cleaning._fill_missing_callsign_for_satellite_waypoints(df_icao_address)
 
     # Construct flight that is diverted
     df_test_4 = df_icao_address[df_icao_address["callsign"] == "BAW506"].copy()
@@ -133,28 +135,30 @@ def test_identify_flight_diversion():
     dt_sec = np.diff(df_test_4["timestamp"].values, append=np.datetime64("NaT")) / np.timedelta64(
         1, "s"
     )
-    flight_phase: adsb.FlightPhaseDetailed = adsb.identify_phase_of_flight_detailed(
+    flight_phase: data_cleaning.FlightPhaseDetailed = jet.identify_phase_of_flight_detailed(
         df_test_4["altitude_baro"].values, dt_sec, threshold_rocd=250, min_cruise_alt_ft=20000
     )
 
     # Identify multiple cruise phase
-    trajectory_check: adsb.TrajectoryCheck = adsb.identify_multiple_cruise_phases_and_diversion(
-        flight_phase,
-        df_test_4["longitude"].values,
-        df_test_4["latitude"].values,
-        df_test_4["altitude_baro"].values,
-        dt_sec,
-        ground_indicator=df_test_4["on_ground"].values,
+    trajectory_check: data_cleaning.TrajectoryCheck = (
+        data_cleaning.identify_multiple_cruise_phases_and_diversion(
+            flight_phase,
+            df_test_4["longitude"].values,
+            df_test_4["latitude"].values,
+            df_test_4["altitude_baro"].values,
+            dt_sec,
+            ground_indicator=df_test_4["on_ground"].values,
+        )
     )
     assert trajectory_check.multiple_cruise_phase is False
     assert trajectory_check.flight_diversion
 
     # Unable to identify unique flights because metadata is the same
-    flights = adsb.separate_unique_flights_from_waypoints(
+    flights = data_cleaning.separate_unique_flights_from_waypoints(
         df_test_4, columns=["tail_number", "aircraft_type_icao", "callsign"]
     )
     assert len(flights) == 1
 
     # It is a flight diversion, so no separation
-    flights = spire.separate_flights_with_multiple_cruise_phase(flights)
+    flights = data_cleaning.separate_flights_with_multiple_cruise_phase(flights)
     assert len(flights) == 1
